code-interpreter-agent/pyproject.toml
[project]
name = "code-interpreter-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonpatch>=1.33",
    "langchain>=0.3.25",
    "langchain-aws>=0.2.23",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langgraph>=0.4.3",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "qrcode>=8.2",
    "tabulate>=0.9.0",
]
